Help the demos find the library and their resources
There are no installation targets provided to install the library in a
well-known location. unix/pkgIndex.tcl contains some code to help scripts
find the library and to perform a renaming dance but this code isn't
available in a well-known location either.

...so this patch simply copy & pastes the snippet from unix/pkgIndex.tcl to
where it is needed to make the demos work.

The existing definition of `dir` prevents the demos from finding their
image resources and isn't compatible with the unix/pkgIndex.tcl library
loading trick so this patch also fixes it up so that it works for me.

Signed-off-by: Andy Bennett <andyjpb@ashurst.eu.org>